Develop

+++ netcal/metrics/confidence/MMCE.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2019-2022 Ruhr West University of Applied Sciences, Bottrop, Germany
+# Copyright (C) 2019-2023 Ruhr West University of Applied Sciences, Bottrop, Germany
 # AND e:fs TechHub GmbH, Gaimersheim, Germany
 #
 # This Source Code Form is subject to the terms of the Apache License 2.0
@@ -142,7 +142,7 @@ class MMCE(object):
             n_samples = float(confidence.size)
 
             # get matched flag and difference
-            matched = (y_batch == labels).astype(np.float)
+            matched = (y_batch == labels).astype(confidence.dtype)
             diff = np.expand_dims(matched - confidence, axis=1)
 
             # now calculate product of differences for each pair
